♻️ feat(new-map-test): Deleting old map and refactoring
