+++ WEB/src/router/index.ts
@@ -6,7 +6,6 @@ import CardView from '@/views/CardView.vue'
 import CardNumberSelector from '@/views/CardNumberSelector.vue'
 import HomeView from '@/views/HomeView.vue'
 import addCardView from '@/views/addCardView.vue'
-import MapViewTest from '@/views/MapViewTest.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -20,11 +19,6 @@ const router = createRouter({
       path: '/map',
       name: 'mainHome',
       component: MapView
-    },
-    {
-      path: '/testMap',
-      name: 'testMap',
-      component: MapViewTest
     },
     {
       path: '/cards',
